requirements.txt
selenium==3.141.0
python-nmap==0.6.1
redis==2.10.6
aliyun-python-sdk-core-v3==2.8.9 
django==2.1.7   